started to separate pieces out of HAAPI
