+++ HomeAssistant/HAAPI.swift
@@ -6,20 +6,19 @@
 //  Copyright © 2016 Robbie Trencheny. All rights reserved.
 //
 
-import Foundation
 import Alamofire
-import PromiseKit
-import CoreLocation
 import AlamofireObjectMapper
-import ObjectMapper
-import DeviceKit
+import PromiseKit
 import Crashlytics
-import UserNotifications
-import RealmSwift
+import CoreLocation
 import CoreMotion
-import Shared
-
+import DeviceKit
+import Foundation
 import KeychainAccess
+import ObjectMapper
+import RealmSwift
+import Shared
+import UserNotifications
 
 // swiftlint:disable file_length
 
@@ -862,383 +861,6 @@ public class HomeAssistantAPI {
                     realm.add(RLMZone(zone: zone), update: true)
                 }
             }
-        }
-    }
-}
-
-class BonjourDelegate: NSObject, NetServiceBrowserDelegate, NetServiceDelegate {
-
-    var resolving = [NetService]()
-    var resolvingDict = [String: NetService]()
-
-    // Browser methods
-
-    func netServiceBrowser(_ netServiceBrowser: NetServiceBrowser,
-                           didFind netService: NetService,
-                           moreComing moreServicesComing: Bool) {
-        NSLog("BonjourDelegate.Browser.didFindService")
-        netService.delegate = self
-        resolvingDict[netService.name] = netService
-        netService.resolve(withTimeout: 0.0)
-    }
-
-    func netServiceDidResolveAddress(_ sender: NetService) {
-        NSLog("BonjourDelegate.Browser.netServiceDidResolveAddress")
-        if let txtRecord = sender.txtRecordData() {
-            let serviceDict = NetService.dictionary(fromTXTRecord: txtRecord)
-            let discoveryInfo = DiscoveryInfoFromDict(locationName: sender.name, netServiceDictionary: serviceDict)
-            NotificationCenter.default.post(name: NSNotification.Name(rawValue: "homeassistant.discovered"),
-                                            object: nil,
-                                            userInfo: discoveryInfo.toJSON())
-        }
-    }
-
-    func netServiceBrowser(_ netServiceBrowser: NetServiceBrowser,
-                           didRemove netService: NetService,
-                           moreComing moreServicesComing: Bool) {
-        NSLog("BonjourDelegate.Browser.didRemoveService")
-        let discoveryInfo: [NSObject: Any] = ["name" as NSObject: netService.name]
-        NotificationCenter.default.post(name: NSNotification.Name(rawValue: "homeassistant.undiscovered"),
-                                        object: nil,
-                                        userInfo: discoveryInfo)
-        resolvingDict.removeValue(forKey: netService.name)
-    }
-
-    private func DiscoveryInfoFromDict(locationName: String,
-                                       netServiceDictionary: [String: Data]) -> DiscoveryInfoResponse {
-        var outputDict: [String: Any] = [:]
-        for (key, value) in netServiceDictionary {
-            outputDict[key] = String(data: value, encoding: .utf8)
-            if outputDict[key] as? String == "true" || outputDict[key] as? String == "false" {
-                if let stringedKey = outputDict[key] as? String {
-                    outputDict[key] = Bool(stringedKey)
-                }
-            }
-        }
-        outputDict["location_name"] = locationName
-        if let baseURL = outputDict["base_url"] as? String {
-            if baseURL.hasSuffix("/") {
-                outputDict["base_url"] = baseURL[..<baseURL.index(before: baseURL.endIndex)]
-            }
-        }
-        return DiscoveryInfoResponse(JSON: outputDict)!
-    }
-}
-
-class Bonjour {
-    var nsb: NetServiceBrowser
-    var nsp: NetService
-    var nsdel: BonjourDelegate?
-
-    init() {
-        let device = Device()
-        self.nsb = NetServiceBrowser()
-        self.nsp = NetService(domain: "local", type: "_hass-ios._tcp.", name: device.name, port: 65535)
-    }
-
-    func buildPublishDict() -> [String: Data] {
-        var publishDict: [String: Data] = [:]
-        if let bundleVersion = Bundle.main.object(forInfoDictionaryKey: "CFBundleVersion") {
-            if let stringedBundleVersion = bundleVersion as? String {
-                if let data = stringedBundleVersion.data(using: .utf8) {
-                    publishDict["buildNumber"] = data
-                }
-            }
-        }
-        if let versionNumber = Bundle.main.object(forInfoDictionaryKey: "CFBundleShortVersionString") {
-            if let stringedVersionNumber = versionNumber as? String {
-                if let data = stringedVersionNumber.data(using: .utf8) {
-                    publishDict["versionNumber"] = data
-                }
-            }
-        }
-        if let permanentID = DeviceUID.uid().data(using: .utf8) {
-            publishDict["permanentID"] = permanentID
-        }
-        if let bundleIdentifier = Bundle.main.bundleIdentifier {
-            if let data = bundleIdentifier.data(using: .utf8) {
-                publishDict["bundleIdentifier"] = data
-            }
-        }
-        return publishDict
-    }
-
-    func startDiscovery() {
-        self.nsdel = BonjourDelegate()
-        nsb.delegate = nsdel
-        nsb.searchForServices(ofType: "_home-assistant._tcp.", inDomain: "local.")
-    }
-
-    func stopDiscovery() {
-        nsb.stop()
-    }
-
-    func startPublish() {
-        //        self.nsdel = BonjourDelegate()
-        //        nsp.delegate = nsdel
-        nsp.setTXTRecord(NetService.data(fromTXTRecord: buildPublishDict()))
-        nsp.publish()
-    }
-
-    func stopPublish() {
-        nsp.stop()
-    }
-
-}
-
-public typealias OnLocationUpdated = ((CLLocation?, Error?) -> Void)
-
-class RegionManager: NSObject {
-
-    let locationManager = CLLocationManager()
-    var backgroundTask: UIBackgroundTaskIdentifier?
-    let activityManager = CMMotionActivityManager()
-    var lastActivity: CMMotionActivity?
-    var lastLocation: CLLocation?
-    var oneShotLocationActive: Bool = false
-
-    var zones: [RLMZone] {
-        let realm = Current.realm()
-        return realm.objects(RLMZone.self).map { $0 }
-    }
-
-    var activeZones: [RLMZone] {
-        let realm = Current.realm()
-        return realm.objects(RLMZone.self).filter(NSPredicate(format: "inRegion == %@",
-                                                              NSNumber(value: true))).map { $0 }
-    }
-
-    internal lazy var coreMotionQueue: OperationQueue = {
-        return OperationQueue()
-    }()
-
-    override init() {
-        super.init()
-        locationManager.desiredAccuracy = kCLLocationAccuracyHundredMeters
-        locationManager.allowsBackgroundLocationUpdates = true
-        locationManager.delegate = self
-        locationManager.distanceFilter = kCLLocationAccuracyHundredMeters
-        startMonitoring()
-        syncMonitoredRegions()
-    }
-
-    private func startMonitoring() {
-        locationManager.startMonitoringSignificantLocationChanges()
-        locationManager.startMonitoringVisits()
-    }
-
-    func triggerRegionEvent(_ manager: CLLocationManager, trigger: LocationUpdateTrigger,
-                            region: CLRegion) {
-        guard let api = HomeAssistantAPI.authenticatedAPI() else {
-           return
-        }
-
-        var trig = trigger
-        guard let zone = zones.filter({ region.identifier == $0.ID }).first else {
-            print("Zone ID \(region.identifier) doesn't exist in Realm, syncing monitored regions now")
-            return syncMonitoredRegions()
-        }
-
-        // Do nothing in case we don't want to trigger an enter event
-        if zone.TrackingEnabled == false {
-            print("Tracking enabled is false")
-            return
-        }
-
-        if zone.IsBeaconRegion {
-            if trigger == .RegionEnter {
-                trig = .BeaconRegionEnter
-            }
-            if trigger == .RegionExit {
-                trig = .BeaconRegionExit
-            }
-        }
-
-        backgroundTask = UIApplication.shared.beginBackgroundTask { [weak self] in
-            self?.endBackgroundTask()
-        }
-
-        let realm = Current.realm()
-        // swiftlint:disable:next force_try
-        try! realm.write {
-            zone.inRegion = (trig == .RegionEnter || trig == .BeaconRegionEnter)
-        }
-
-        print("Submit location for zone \(zone.ID) with trigger \(trig.rawValue)")
-
-        api.submitLocation(updateType: trig, location: nil, visit: nil, zone: zone)
-    }
-
-    func startMonitoring(zone: RLMZone) {
-        if let region = zone.region() {
-            locationManager.startMonitoring(for: region)
-        }
-
-        activityManager.startActivityUpdates(to: coreMotionQueue) { activity in
-            self.lastActivity = activity
-        }
-    }
-
-    @objc func syncMonitoredRegions() {
-        // stop monitoring for all regions
-        locationManager.monitoredRegions.forEach { region in
-            print("Stopping monitoring of region \(region.identifier)")
-            locationManager.stopMonitoring(for: region)
-        }
-
-        // start monitoring for all existing regions
-        zones.forEach { zone in
-            print("Starting monitoring of zone \(zone)")
-            startMonitoring(zone: zone)
-        }
-    }
-
-    func checkIfInsideAnyRegions(location: CLLocationCoordinate2D) -> Set<CLRegion> {
-        return self.locationManager.monitoredRegions.filter { (region) -> Bool in
-            if let circRegion = region as? CLCircularRegion {
-                // print("Checking", circRegion.identifier)
-                return circRegion.contains(location)
-            }
-            return false
-        }
-    }
-}
-
-// MARK: CLLocationManagerDelegate
-extension RegionManager: CLLocationManagerDelegate {
-    func locationManager(_ manager: CLLocationManager, didChangeAuthorization status: CLAuthorizationStatus) {
-        if status == .authorizedAlways {
-            prefs.setValue(true, forKey: "locationEnabled")
-            prefs.synchronize()
-        }
-    }
-
-    func locationManager(_ manager: CLLocationManager, didUpdateLocations locations: [CLLocation]) {
-        guard let api = HomeAssistantAPI.authenticatedAPI() else {
-            return
-        }
-        if self.oneShotLocationActive {
-            print("NOT accepting region manager update as one shot location service is active")
-            return
-        }
-        print("RegionManager: Got location, stopping updates!", locations.last.debugDescription, locations.count)
-        api.submitLocation(updateType: .SignificantLocationUpdate, location: locations.last, visit: nil,
-                           zone: nil)
-
-        self.lastLocation = locations.last
-
-        locationManager.stopUpdatingLocation()
-    }
-
-    func locationManager(_ manager: CLLocationManager, didEnterRegion region: CLRegion) {
-        print("Region entered", region.identifier)
-        triggerRegionEvent(manager, trigger: .RegionEnter, region: region)
-    }
-
-    func locationManager(_ manager: CLLocationManager, didExitRegion region: CLRegion) {
-        print("Region exited", region.identifier)
-        triggerRegionEvent(manager, trigger: .RegionExit, region: region)
-    }
-
-    func locationManager(_ manager: CLLocationManager, didVisit visit: CLVisit) {
-        guard let api = HomeAssistantAPI.authenticatedAPI() else {
-            return
-        }
-
-        print("Visit logged")
-        api.submitLocation(updateType: .Visit, location: nil, visit: visit, zone: nil)
-    }
-
-    func locationManager(_ manager: CLLocationManager, didFailWithError error: Error) {
-        if let clErr = error as? CLError {
-            let realm = Current.realm()
-            // swiftlint:disable:next force_try
-            try! realm.write {
-                let locErr = LocationError(err: clErr)
-                realm.add(locErr)
-            }
-            print(clErr.debugDescription)
-            if clErr.code == CLError.locationUnknown {
-                // locationUnknown just means that GPS may be taking an extra moment, so don't throw an error.
-                return
-            }
-        } else {
-            print("other error:", error.localizedDescription)
-        }
-    }
-
-    func locationManager(_ manager: CLLocationManager, didStartMonitoringFor region: CLRegion) {
-//        let insideRegions = checkIfInsideAnyRegions(location: lastLocation.coordinate)
-//        for inside in insideRegions {
-//            print("System reports inside for zone", inside.identifier)
-//        }
-        print("Started monitoring region", region.identifier)
-        locationManager.requestState(for: region)
-    }
-
-    func locationManager(_ manager: CLLocationManager, didDetermineState state: CLRegionState, for region: CLRegion) {
-        var strState = "Unknown"
-        if state == .inside {
-            strState = "Inside"
-        } else if state == .outside {
-            strState = "Outside"
-        } else if state == .unknown {
-            strState = "Unknown"
-        }
-        print("\(strState) region", region.identifier)
-    }
-}
-
-// MARK: BackgroundTask
-extension RegionManager {
-    func endBackgroundTask() {
-        if backgroundTask != UIBackgroundTaskInvalid {
-            UIApplication.shared.endBackgroundTask(backgroundTask!)
-            backgroundTask = UIBackgroundTaskInvalid
-        }
-    }
-}
-
-class LocationManager: NSObject {
-    let locationManager = CLLocationManager()
-    var onLocationUpdated: OnLocationUpdated
-
-    init(onLocation: @escaping OnLocationUpdated) {
-        onLocationUpdated = onLocation
-        super.init()
-        locationManager.desiredAccuracy = kCLLocationAccuracyHundredMeters
-        locationManager.allowsBackgroundLocationUpdates = true
-        locationManager.distanceFilter = kCLLocationAccuracyHundredMeters
-        locationManager.delegate = self
-        locationManager.startUpdatingLocation()
-    }
-}
-
-extension LocationManager: CLLocationManagerDelegate {
-    func locationManager(_ manager: CLLocationManager, didUpdateLocations locations: [CLLocation]) {
-        print("LocationManager: Got location, stopping updates!", locations.last.debugDescription, locations.count)
-        onLocationUpdated(locations.first, nil)
-        manager.stopUpdatingLocation()
-        manager.delegate = nil
-    }
-
-    func locationManager(_ manager: CLLocationManager, didFailWithError error: Error) {
-        if let clErr = error as? CLError {
-            let realm = Current.realm()
-            // swiftlint:disable:next force_try
-            try! realm.write {
-                let locErr = LocationError(err: clErr)
-                realm.add(locErr)
-            }
-            print(clErr.debugDescription)
-            if clErr.code == CLError.locationUnknown {
-                // locationUnknown just means that GPS may be taking an extra moment, so don't throw an error.
-                return
-            }
-            onLocationUpdated(nil, clErr)
-        } else {
-            print("other error:", error.localizedDescription)
-            onLocationUpdated(nil, error)
         }
     }
 }
