+++ HomeAssistant/Location/LocationManager.swift
@@ -1,0 +1,55 @@
+//
+//  LocationManager.swift
+//  HomeAssistant
+//
+//  Created by Stephan Vanterpool on 8/24/18.
+//  Copyright © 2018 Robbie Trencheny. All rights reserved.
+//
+
+import CoreLocation
+import Foundation
+import Shared
+
+class LocationManager: NSObject {
+    let locationManager = CLLocationManager()
+    var onLocationUpdated: OnLocationUpdated
+
+    init(onLocation: @escaping OnLocationUpdated) {
+        onLocationUpdated = onLocation
+        super.init()
+        locationManager.desiredAccuracy = kCLLocationAccuracyHundredMeters
+        locationManager.allowsBackgroundLocationUpdates = true
+        locationManager.distanceFilter = kCLLocationAccuracyHundredMeters
+        locationManager.delegate = self
+        locationManager.startUpdatingLocation()
+    }
+}
+
+extension LocationManager: CLLocationManagerDelegate {
+    func locationManager(_ manager: CLLocationManager, didUpdateLocations locations: [CLLocation]) {
+        print("LocationManager: Got location, stopping updates!", locations.last.debugDescription, locations.count)
+        onLocationUpdated(locations.first, nil)
+        manager.stopUpdatingLocation()
+        manager.delegate = nil
+    }
+
+    func locationManager(_ manager: CLLocationManager, didFailWithError error: Error) {
+        if let clErr = error as? CLError {
+            let realm = Current.realm()
+            // swiftlint:disable:next force_try
+            try! realm.write {
+                let locErr = LocationError(err: clErr)
+                realm.add(locErr)
+            }
+            print(clErr.debugDescription)
+            if clErr.code == CLError.locationUnknown {
+                // locationUnknown just means that GPS may be taking an extra moment, so don't throw an error.
+                return
+            }
+            onLocationUpdated(nil, clErr)
+        } else {
+            print("other error:", error.localizedDescription)
+            onLocationUpdated(nil, error)
+        }
+    }
+}
