+++ Podfile
@@ -31,6 +31,8 @@ end
 target 'Shared' do
     pod 'Alamofire', '4.7.3'
     pod 'AlamofireObjectMapper', '5.1.0'
+    pod 'Crashlytics', '3.10.2'
+    pod 'DeviceKit', '1.7.0'
     pod 'KeychainAccess', '3.1.1'
     pod 'ObjectMapper', '3.3.0'
     pod 'PromiseKit', '6.3.0'
